refactor: split openai responses stream adapter
